Rewrite most things
ok where do I even begin, firstly, all the positional and vector stuff has been packed up into the Position and Vector classes, which I something I wanted to do a while ago but couldn't be bothered to, as it would require a significant rewrite.
the blocking attribute is now called "solid"
I have rewritten the Elements class (somewhat from scratch), and I have incorporated the new Position and Vector classes, as well as redoing the rect method to return two Positions rather than a numpy linspace.
Lots of things have changed, but overall the actual main.py functions need to be rewritten for the new format, and the template functions need to be completed and cleaned up