Walls, ect
ADDED WALLS it was a lot easier than I expected, I guess that's because I redesigned collision and placement so extensively that it was easy to slide in. Walls longer than the player must be at least (player speed) thick to always work, since thinner walls can be clipped through by the player's corners not being in the wall dimensions. I now plan to make walls toggleable, and to add recharging back to flippers.