Boosters
Added boosters, which allow movement in the direction they face. The player is boosted in that direction by the magnitude of the boost multiplied by the player's movement speed. The last piece of the puzzle I need now is the buttons, and then I will get to assembly and (probably not) testing.